# server/openapi/simpasid.v1.yaml
openapi: 3.1.0
info:
  title: API аккаунта СИМПАС
  version: "1.0.0"
  description: |
    Единый API личности СИМПАС. Этим же API работает портал аккаунта и мобильные
    приложения: отдельного внутреннего пути не существует. Если чего-то нет здесь —
    этого не может быть и на экране.

    Чего в этом API нет и не появится: пароля, телефона, SMS, ролей, тарифов,
    признака «психолог», аватара и ФИО из внешнего сервиса. Продукт узнаёт права
    из своей базы по `sub`.
  contact:
    name: Поддержка СИМПАС
    url: https://auth.cmpas.ru/dev
servers:
  - url: https://auth.cmpas.ru/v1
    description: Боевой контур
...
security:
  - bearerAuth: []
paths:
  /account:
    get:
      operationId: getAccount
      tags: [account]
      summary: Профиль
      description: Профиль текущей учётной записи и перечень подключённых продуктов.
      responses:
        "200":
          description: Профиль
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Account" }
        "401": { $ref: "#/components/responses/Unauthorized" }
    patch:
      operationId: updateAccount
      tags: [account]
      summary: Изменить имя
      description: Меняет имя, которое человек видит в письмах и в портале. Профиль продукта здесь не редактируется.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AccountPatch" }
      responses:
        "200":
          description: Обновлённый профиль
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Account" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Ключ доступа, выданный при входе.

    firstPartyClient:
      type: apiKey
      in: header
      name: X-Client-Id
      description: |
        Идентификатор приложения. Первичный токен-API открыт ТОЛЬКО
        приложениям экосистемы (first_party); всем остальным — вход по
        OIDC с редиректом. Запрос без этого заголовка или с чужим
        идентификатором получает 403 forbidden_client.

        Секрета здесь нет и быть не может: приложение на устройстве
        человека сохранить секрет не в состоянии, а вход спроектирован
        так, что знание идентификатора ничего не даёт.
...
  responses:
    Unauthorized:
      description: Ключ доступа не предъявлен или не действует
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    BadRequest:
      description: Запрос не разобран
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
  schemas:
    ProductCode:
      type: string
      # Набор РАСТЁТ: сегодня добавились ШАГИ, и они вряд ли последние.
      # Значение, дописанное в закрытый enum ответа, ломает клиента,
      # написанного по прежней спецификации: такого продукта он не ждёт.
      # x-extensible-enum говорит честно — неизвестное значение обязано
      # быть обработано. В запросах ProductCode не используется ни разу,
      # так что строгость приёма не теряется: её держат перечень в коде
      # и ограничение схемы в базе.
      x-extensible-enum: [practice, zapiski, moments, steps]
      description: Продукт экосистемы.
...
    Account:
      type: object
      description: Учётная запись СИМПАС.
      required: [id, email, email_verified, products]
      properties:
        id:
          type: string
          description: Устойчивый идентификатор учётной записи.
          examples: ["acc_8f21a4c0"]
        email:
          type: string
          description: Основная почта.
          examples: ["t@ya.ru"]
        email_verified:
          type: boolean
          description: Подтверждена ли основная почта.
          examples: [true]
        display_name:
          type: [string, "null"]
          description: Имя, которое человек ввёл сам.
          examples: ["Илья"]
        products:
          type: array
          description: Подключённые продукты.
          items: { $ref: "#/components/schemas/ProductCode" }
          examples: [["practice", "zapiski"]]
    AccountPatch:
      type: object
      description: Изменяемые поля профиля.
      properties:
        display_name:
          type: [string, "null"]
          maxLength: 100
          description: Имя. Пустое значение убирает имя.
          examples: ["Илья"]
...
    Error:
      type: object
      description: Отказ.
      required: [error]
      properties:
        error:
          type: string
          description: Машинный код отказа. Текст для человека берётся из реестра формулировок, а не отсюда.
          examples: ["last_login_method"]
...